Align people countdown conversion scheduling
Carry the state-five frame count as a float immediately after retrieval so MWCC emits the target conversion and division schedule.

Constraint: Genuine C only; preserve the recovered state machine semantics.
Rejected: Changing the visibility flag type | produced no focused score change.
Confidence: high
Scope-risk: narrow
Directive: Keep frameCount in floating form while tuning remaining state-five allocation.
Tested: ninja build/GC6E01/src/game/people/people.o; focused fn_80181850 91.05147% -> 93.460785%; git diff --check.
Not-tested: Runtime countdown precision beyond target-equivalent conversion.

+++ src/game/people/people.c
@@ -562,7 +562,7 @@ void fn_80181850(void)
     u8 layer;
     u8 subLayer;
     s32 shadowCount;
-    s32 frameCount;
+    f32 frameCount;
     u32 ticks;
     f32 angle;
     f32 fullTurn;
@@ -607,10 +607,10 @@ void fn_80181850(void)
                 switch (entry->subState) {
                 case 0:
                     if (entry->animBlendFactor > lbl_8047D7A0) {
-                        frameCount = fn_800D37CC();
+                        frameCount = (f32)fn_800D37CC();
                         ticks = fn_800D3088();
                         entry->animBlendFactor -=
-                            (f32)ticks / (f32)frameCount;
+                            (f32)ticks / frameCount;
                         if (entry->animBlendFactor < lbl_8047D7A0) {
                             entry->animBlendFactor = lbl_8047D7A0;
                         }
